+++ docs/homepage/source/conf.py
@@ -1,0 +1,35 @@
+# Configuration file for the Sphinx documentation builder.
+#
+# For the full list of built-in configuration values, see the documentation:
+# https://www.sphinx-doc.org/en/master/usage/configuration.html
+
+# -- Project information -----------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
+# 
+# Must match what is in pyproject.toml, also update proposed_release accordingly
+# when generating the docs for an actual, not proposed, release.
+#
+
+project = 'Boring Math'
+copyright = '2016-2025, Geoffrey R. Scheller'
+author = 'Geoffrey R. Scheller'
+release = '2.0.0' 
+
+# -- General configuration ---------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
+
+extensions = [
+    'sphinx.ext.autodoc',
+]
+
+autodoc_member_order = 'bysource'
+autoclass_content = 'both'
+
+templates_path = ['_templates']
+exclude_patterns: list[str] = []
+
+# -- Options for HTML output -------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
+
+html_theme = 'piccolo_theme'
+html_static_path = ['_static']