+++ src/boring_math/homepage/__init__.py
@@ -1,0 +1,30 @@
+"""
+The boring-math PyPI project is the central component of my
+*Boring Math* hobby projects. It has several purposes.
+
+- claims the boring-math name on PyPI for the overall effort
+- hosts the effort's homepage on *GitHub Pages*
+- hosts detailed Sphinx documentation for each *Boring Math* project
+- contains the ``boring_math.homepage`` module
+
+  - which is used in the construction of the homepage
+  - keeps important information available and modifiable
+
+    - outside of Sphinx boilerplate
+
+- provides locations for effort's overall CHANGELOG and README files
+
+  - the README.md file provides
+
+    - the boring-math PyPI description giving an overview of *Boring Math*
+    - provide links to the individual projects on PyPI and their GitHub repos
+
+  - each *Boring Math* project has its on CHANGELOG and README files
+
+There is no need to install boring-math itself, but does not hurt
+anything if it is installed.
+"""
+
+__author__ = 'Geoffrey R. Scheller'
+__copyright__ = 'Copyright (c) 2016-2025 Geoffrey R. Scheller'
+__license__ = 'Apache License 2.0'