+++ src/boring_math/homepage/development.py
@@ -1,0 +1,23 @@
+"""
+Semantic versioning
+-------------------
+
+Maintainer has adopted strict 3 digit
+`semantic versioning <https://semver.org>`_
+and does not use
+`caps on dependencies <https://iscinumpy.dev/post/bound-version-constraints>`_.
+This allows for more package management flexibility for developers and
+access to the latest features.
+
+Periodically known consistent releases of versions are done for those
+concerned with stability. These are also posted in the project's CHANGELOG.
+
+CHANGELOG
+---------
+
+Pythonic FP overarching
+`CHANGELOG <https://github.com/grscheller/boring-math/blob/main/CHANGELOG.rst>`_.
+
+Each individual Boring Math project also has its own CHANGELOG.
+
+"""