+++ src/boring_math/homepage/intro.py
@@ -1,0 +1,15 @@
+"""
+Daddy's Boring Math Projects
+============================
+
+Collection of maintainers mathematical hobby projects.
+
+The overall project's name is **Pythonic FP** and consists of PyPI
+projects with PyPI and GitHub names beginning `boring-math`. All
+projects are installed under the Python ``boring_math`` namespace.
+
+The name "Daddy's Boring Math Projects" was suggested by my then
+13 year old daughter Mary when I asked her for a good name for
+this effort.
+
+"""